add useExportToPng
